+++ src/RedisStream.php
@@ -20,7 +20,28 @@ class RedisStream
         // do nothing
     }
 
-    public function xadd(string $key, string $id = '*', array $data): string
+    /**
+     * Function to add new data to the stream
+     *
+     * @param string $key   The stream key
+     * @param string $id    New object ID
+     * @param array $data   The data that will be input. $data doesn't support a nested array
+     * @return string
+     */
+    public static function xadd(string $key, string $id = '*', array $data): string
+    {
+        return (new static )->newXadd($key, $id, $data);
+    }
+
+    /**
+     * Function to create static function 'xadd'
+     *
+     * @param string $key   The stream key
+     * @param string $id    New object ID
+     * @param array $data   The data that will be input. $data doesn't support a nested array
+     * @return string
+     */
+    public function newXadd(string $key, string $id = '*', array $data): string
     {
         $fieldsAndValues = $this->parseData($data);
 
@@ -31,23 +52,38 @@ class RedisStream
             ...$fieldsAndValues,
         ]);
 
-        if ($result !== 'OK') {
-            throw new \Exception($result);
-        }
-
         return $result;
     }
 
-    public function xread(array $streams, array $ids, array $options): Collection
+    /**
+     * Function to read from the stream
+     *
+     * @param array $streams    Streams that will be read
+     * @param array $ids        IDs from the stream that will retrieve
+     * @param array $options
+     * @return \Illuminate\Support\Collection
+     */
+    public static function xread(array $streams, array $ids, array $options): Collection
+    {
+        return (new static )->xread($streams, $ids, $options);
+    }
+
+    /**
+     * Function to create static function 'xread'
+     *
+     * @param array $streams    Streams that will be read
+     * @param array $ids        IDs from the stream that will retrieve
+     * @param array $options
+     * @return array
+     */
+    public function newXread(array $streams, array $ids, array $options): array
     {
         $redisArguments = [
             Commands::XREAD,
             ...$options,
             Options::OPTION_STREAMS,
+            ...$streams,
         ];
-        foreach ($streams as $stream) {
-            $redisArguments[] = $stream;
-        }
         foreach ($ids as $id) {
             $redisArguments[] = !$id ? '0' : $id;
         }
@@ -58,12 +94,20 @@ class RedisStream
             throw new \Exception($results);
         }
 
-        return $this->parseResults($results);
+        return $this->parseResults($results ?? []);
     }
 
-    public function xdel(string $key, array $ids): void
+    /**
+     * Function to delete object(s) from stream
+     *
+     * @param string $key   Stream that will be read
+     * @param array $ids    A List of id that will be delete
+     * @return void
+     */
+    public static function xdel(string $key, array $ids): void
     {
         $result = Redis::executeRaw([
+            Commands::XDEL,
             $key,
             ...$ids,
         ]);
@@ -73,9 +117,18 @@ class RedisStream
         }
     }
 
-    public function xack(string $key, string $group, array $ids): void
+    /**
+     * Function to acknowledge event(s)
+     *
+     * @param string $key       The stream that will be read
+     * @param string $group     The consumer group
+     * @param array $ids        ID of objects
+     * @return void
+     */
+    public static function xack(string $key, string $group, array $ids): void
     {
         $result = Redis::executeRaw([
+            Commands::XACK,
             $key,
             $group,
             ...$ids,
@@ -86,9 +139,20 @@ class RedisStream
         }
     }
 
-    public function xgroup(string $option, string $key, string $groupname, bool $mkstream = false, array $arguments = []): void
+    /**
+     * Function to create stream group
+     *
+     * @param string $option        XGROUP Options as listed on \Afikrim\LaravelRedisStream\Data\XGROUPOptions
+     * @param string $key           The stream that will be read
+     * @param string $groupname     The group name
+     * @param boolean $mkstream     To make a stream when creating group
+     * @param array $arguments      Extra arguments, like ID or consumername. eg, ['$'] || ['consumername']
+     * @return void
+     */
+    public static function xgroup(string $option, string $key, string $groupname, bool $mkstream = false, array $arguments = []): void
     {
         $redisArguments = [
+            Commands::XGROUP,
             $option,
             $key,
             $groupname,
@@ -101,14 +165,40 @@ class RedisStream
 
         $result = Redis::executeRaw($redisArguments);
 
-        if ($result !== 'OK') {
+        if ($result !== 'OK' && !preg_match('/^\d+$/', $result)) {
             throw new \Exception($result);
         }
     }
 
-    public function xreadgroup(string $group, string $consumer, array $streams, array $ids, array $options)
+    /**
+     * Function to read from a certain group
+     *
+     * @param string $group         The group
+     * @param string $consumer      Consumer name
+     * @param array $streams        Streams that will be read
+     * @param array $ids            ID that will be retrieve
+     * @param array $options        Extra options, as define in \Afikrim\LaravelRedisStream\Data\Options. eg, [Options::OPTION_BLOCK, 2000]
+     * @return array
+     */
+    public static function xreadgroup(string $group, string $consumer, array $streams, array $ids, array $options): array
+    {
+        return (new static )->newXreadgroup($group, $consumer, $streams, $ids, $options);
+    }
+
+    /**
+     * Function to read from a certain group
+     *
+     * @param string $group         The group
+     * @param string $consumer      Consumer name
+     * @param array $streams        Streams that will be read
+     * @param array $ids            ID that will be retrieve
+     * @param array $options        Extra options, as define in \Afikrim\LaravelRedisStream\Data\Options. eg, [Options::OPTION_BLOCK, 2000]
+     * @return array
+     */
+    public function newXreadgroup(string $group, string $consumer, array $streams, array $ids, array $options): array
     {
         $redisArguments = [
+            Commands::XREADGROUP,
             Options::OPTION_GROUP,
             $group,
             $consumer,
@@ -128,10 +218,16 @@ class RedisStream
             throw new \Exception($results);
         }
 
-        return $this->parseResults($results);
+        return $this->parseResults($results ?? []);
     }
 
-    private function parseData(array $raw)
+    /**
+     * Function to parse inserted data to redis arguments
+     *
+     * @param array $raw
+     * @return array
+     */
+    protected function parseData(array $raw): array
     {
         $data = [];
         foreach ($raw as $field => $value) {
@@ -142,7 +238,13 @@ class RedisStream
         return $data;
     }
 
-    private function parseResult(array $result)
+    /**
+     * Function to parse read result to associative array
+     *
+     * @param array $result
+     * @return array
+     */
+    protected function parseResult(array $result): array
     {
         [$key, $rawData] = $result;
 
@@ -158,20 +260,24 @@ class RedisStream
                 return ['id' => $id, 'data' => $data];
             });
 
-        return ['key' => $key, 'data' => $data];
+        return ['key' => $key, 'data' => $data->toArray()];
     }
 
-    private function parseResults(array $results)
+    /**
+     * Function to parse read result to associative array
+     *
+     * @param array $results
+     * @return array
+     */
+    protected function parseResults(array $results): array
     {
         $data = collect($results)
             ->reduce(function ($prev, $current) {
                 $result = $this->parseResult($current);
 
-                return array_merge(
-                    [$result], $prev
-                );
+                return [$result, ...$prev];
             }, []);
 
-        return collect($data);
+        return (array) $data;
     }
 }